Improve ghcid experience
The .ghci config loads the src dir containing all the library code. The
idea of this is that until ghci supports multiple targets we need to
load the library code along side the test code ... this is so that we
can then reload all src modules and run the test suite when they
compile.

In order to do this we need to have all the library's deps loaded for
the test module also, we'll also need all the language extensions. To
address this I'm planning to add HPack to make writing this cabal file a
little less onerous.