Don't accept nixpkgs as arg to shell.nix
Changing nix packages is a bit of a commitment. I'm going add a script
to make it easier to switch version, but in such a way that once you've
got the version you need you can just `git commit` and you're right to
go. I'd hate for someone to find a working version and somehow not be
able to find a reference to it to put in nixpkgs.json etc.