Travis: remove nix store caching and avoid containers
So build times improved a lot by following the advice from nixos.org
(see commit where local caching was added) though it appears that is
because of the sudo, rather than the cache.

Being that the cache command fails and that asking for sudo, which moves
the build onto beefier machines, I'm going to delete the caching and
just ask for sudo.

This does simplify things abit, I've not seen a build which has taken
advantage of the local cache yet and times appear to be around 10
minutes. Which might be good enough for now.