Remove Buildkite configuration
In order to run a Buildkite agent I need a server up and running
somewhere. At the momement that's a Digital Ocean 2 cpu + 4g ram Droplet
in the San Francisco region. That's $20 a month, so it was a good
experiement to set it up, but for now I'd like something that anyone
else can easily spin up if they want. Also it'd be nice if third-party
PRs can have the suite run.

The history will still be here so if there's ever a desire to go back to
Buildkite (they're talking about adding read only access to pipeline)
it might be as easy as reverting this commit.