Add busybox to docker image, default tag to latest
For the tag:
It appears dockerTools doesn't like the sort of tagging that heroku
requires. So I'm letting it default to "latest" and then using the docker
cli to tag it following heroku's naming convention.

For the busybox addition:
Heroku _really_ wants to start the app using `/bin/sh -c` ... so this is
a stop-gap until I figure out how to make it happy _only_ running the
app binary.