The Final Push
This gets the postman collection from the Realworld repo passing. Well,
all except for the auth header change from "Token" to "Bearer".

I was pretty surprised at how easy it was to add the following bool to
Profiles. When I saw I'd missed it I was pretty worried that it'd be a
lot of mucking about, but no. All good.

I have no idea what's going on with this big join'd queries I'm doing,
something that might be good to do now would be to run EXPLAIN on the
debug output of the feed / all articles queries. We'll see how
enthusiastic I get.